Completed Acceptance Testing

+++ src/main/java/com/dennis/oconnell/samplejavaapi/seeder/TimesheetDbSeeder.java
@@ -27,6 +27,9 @@ public class TimesheetDbSeeder implements InitializingBean {
 
     public void Seed() throws Exception {
 
+        if(timesheetRepository.count() > 0){
+            return;
+        }
 
         // reset table
         timesheetRepository.deleteAll();
